% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{check_args_match}
\alias{check_args_match}
\title{Checks that arguments passed matched that in function}
\usage{
check_args_match(callback, ...)
}
\arguments{
\item{callback}{Callback function to check against.}

\item{...}{Arguments.}
}
\description{
Checks that arguments passed matched that in function
}
\keyword{.Internal}
