% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{constructors}
\alias{constructors}
\alias{construct_defaults_fn}
\alias{construct_converters_fn}
\title{Constructors}
\usage{
construct_defaults_fn(fn)

construct_converters_fn(fn)
}
\description{
Adds classes to returned function to allow print method.
}
\keyword{internal}
