% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{converters.}
\alias{converters.}
\alias{as_character}
\alias{as_integer}
\alias{as_numeric}
\alias{as_date}
\alias{as_posixct}
\alias{as_posixlt}
\alias{as_dataframe}
\alias{as_list}
\title{Type Converters}
\format{
An object of class \code{list} of length 2.

An object of class \code{list} of length 2.

An object of class \code{list} of length 2.

An object of class \code{list} of length 2.

An object of class \code{list} of length 2.

An object of class \code{list} of length 2.

An object of class \code{list} of length 2.

An object of class \code{list} of length 2.
}
\usage{
as_character

as_integer

as_numeric

as_date

as_posixct

as_posixlt

as_dataframe

as_list
}
\description{
Type converters for convenience.
}
\keyword{datasets}
