% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{create_type}
\alias{create_type}
\title{Create type}
\usage{
create_type(r_converter, js_checker = NULL, name = NULL)
}
\arguments{
\item{r_converter}{R converter function to transform the value.
received from the server to the desired type in R.}

\item{js_checker}{JavaScript checker function to check if the value
received from the client is of the desired type.}

\item{name}{Name of the type.}
}
\description{
Create a new type with its associated checker and converter.
}
