% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{get_args}
\alias{get_args}
\title{Get arguments from function}
\usage{
get_args(handler, schemas)
}
\arguments{
\item{handler}{Function handler.}

\item{schemas}{List of schemas.}
}
\description{
Get arguments from handler.
}
\keyword{internal}
