% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{http_response_json}
\alias{http_response_json}
\title{HTTP response JSON}
\usage{
http_response_json(body, status = 200)
}
\arguments{
\item{body}{Body of response.}

\item{status}{HTTP status code.}
}
\description{
Sends a JSON response.
}
\keyword{internal}
