% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{parse_args}
\alias{parse_args}
\title{Parse Arguments}
\usage{
parse_args(args = list(), converters = list())
}
\arguments{
\item{args}{Parsed query string.}

\item{converters}{List of converters.}
}
\description{
Attempts to parse query string for embeds into R objects.
}
\keyword{internal}
