% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-string.R
\name{parse_query_string}
\alias{parse_query_string}
\title{Parse a query string into a list}
\usage{
parse_query_string(str)
}
\arguments{
\item{str}{Query string.}
}
\description{
Parses the query string from the request into a list.
}
\keyword{internal}
