% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{converters.}
\alias{converters.}
\alias{Character}
\alias{Integer}
\alias{Numeric}
\alias{Date}
\alias{Posixct}
\alias{Posixlt}
\alias{Dataframe}
\alias{List}
\title{Type Converters}
\format{
An object of class \code{list} of length 2.

An object of class \code{list} of length 2.

An object of class \code{list} of length 2.

An object of class \code{list} of length 2.

An object of class \code{list} of length 2.

An object of class \code{list} of length 2.

An object of class \code{list} of length 2.

An object of class \code{list} of length 2.
}
\usage{
Character

Integer

Numeric

Date

Posixct

Posixlt

Dataframe

List
}
\description{
Type converters for convenience.
}
\keyword{datasets}
